rename: PopulationCompResponseの一部がCompoになっていてファイル名と定義名と一致していなかったのでファイル名を変更した
