+++ src/types/api/models/populationComp/PopulationCompResponse.ts
